// backend/src/common/types/file.multer.d.ts
export type FileMulter = {
  fieldname: string;
  originalname: string;
  encoding: string;
  mimetype: string;
  destination: string;
  filename: string;
  path: string;
  size: number;
};

export type Arquivo = {
  uuid: string;
  url: string;
  destination: string;
  encoding: string;
  filename: string;
  mimetype: string;
  originalname: string;
  path: string;
  size: number;
};